Rename helpers_w_NUMBA to helpers_numba
It's not recommended to include capitals in python module names. Some filesystems are not case-sensitive, or even case-preserving, so they can become mangled to lowercase automatically.
